refactor(mojo): extract expr/value/lib modules + emit-context seam
Decompose the monolithic mojo-adapter.ts into go-template-style domain
modules (issue #2018 track D). Pure free functions and the ParsedExpr
emitters move out behind a narrow `MojoEmitContext` seam; the main
adapter keeps its recursive lowering but delegates the extracted
clusters. Behaviour is unchanged — adapter-tests conformance stays
byte-identical (464 conformance cases, full suite 527 pass / 0 fail).

New layout under packages/adapter-mojolicious/src/adapter/:
  emit-context.ts        MojoEmitContext interface (the seam)
  lib/types.ts           PrimitiveSpec, MojoRenderCtx, MojoAdapterOptions
  lib/constants.ts       MOJO_TEMPLATE_PRIMITIVES + emit map
  lib/perl-naming.ts     perlHashKey, perlIdentifierFromMarkerId
  lib/ir-scope.ts        resolveJsxChildrenProp, collectRootScopeNodes,
                         referencedVarsAreAvailable
  value/parsed-literal.ts parsePureStringLiteral + string-type helpers
  expr/operand.ts        isStringTypedOperand, emitIndexAccessPerl
  expr/array-method.ts   renderArrayMethod/Sort/Reduce/Flat/FlatMap
  expr/emitters.ts       MojoFilterEmitter, MojoTopLevelEmitter
